+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Fixed
+
+- Invisible postal code invalidation when editing an address.
+
 ## [1.10.4] - 2019-10-23
 
 ### Fixed